convert OrderWithinStage from int to short
We'll never need the full range of integers.
Converting to short will allow us to represent
stages as a range within int if that becomes
more useful down the road.

Also added a few more MiddlewareStep instances
and starting all steps at 0 instead of MinValue.

I think it's easier to reason about.

+++ CommandDotNet.Tests/FeatureTests/ParseDirective/ParseDirectiveVerifier.cs
@@ -83,13 +83,13 @@ namespace CommandDotNet.Tests.FeatureTests.ParseDirective
                 appRunner.Configure(c =>
                     c.UseMiddleware((ctx, next) => 
                             throw new Exception("throwAtInvoke exception"),
-                        MiddlewareStages.Invoke, int.MinValue));
+                        MiddlewareStages.Invoke, short.MinValue));
             }
 
             appRunner.Configure(c => 
                 c.UseMiddleware((ctx, next) => 
                         throw new Exception("parse should exit before method invocation"),
-                MiddlewareStages.Invoke, int.MaxValue-1));
+                MiddlewareStages.Invoke, short.MaxValue-1));
 
             var result = appRunner
                 .UseParseDirective()
